Bug 2043518: set degraded and available status based on Prometheus pod status
NOTE: This patch addresses 2 related Bugs:
   1. Bug 2043518: Better message in the CMO degraded/unavailable conditions when pods can't be scheduled
   2. Bug 2039411: Monitoring operator reports unavailable=true while one Prometheus pod is ready

The patch sets the Degraded and Availability status of ClusterOperator
by inspecting Prometheus Resource status. To achieve this, a new
PrometheusValidationTask is added that inspects the Prometheus resource
status and returns degraded and/or unavailable validation errors.

If any of the pods are running, the state will be marked as `Degraded`
but `Available` will still be true as opposed to the previous
implementation which incorrectly reported `Availability` as `False`.

Thus, given the following invalid monitoring config which uses a non-existent PVC

```yaml
  prometheusK8s:
    volumeClaimTemplate:
      metadata:
        name: prometheus-db
      spec:
        storageClassName: foo
        resources: { requests: {storage: 100Mi} }
```

It now correctly marks the ClusterOperator status as follows (formatted
for readability)

```
 {
    "type": "Degraded"
    "message": "none of the prometheus are running",
    "reason": "UpdatingPrometheusK8S",
    "status": "True",
  },
  {
    "type": "Available"
    "message": "prometheus pod openshift-monitoring/prometheus-k8s-0 is pending
               due to Unschedulable
               : 0/1 nodes are available
               : 1 pod has unbound immediate PersistentVolumeClaims." ,
    "reason": "UpdatingPrometheusK8S",
    "status": "False",
  }
]
```

Signed-off-by: Sunil Thaha <sthaha@redhat.com>